Simple app to demonstrate signing and completing an activity.
The main app component is initialized with the Bundle args from the Android intent as its initial props. The JS app looks for a prop, "content", and puts the value in the TextInput field. The user can then edit the value, and click "Sign", which will call `setResult` and `finish` on the Android activity, as well as put the JSON object in the bundle to send back to the calling app.